move wait logic to separate function
