+++ ydb/aio/_utilities.py
@@ -1,3 +1,6 @@
+import asyncio
+
+
 class AsyncResponseIterator(object):
     def __init__(self, it, wrapper):
         self.it = it.__aiter__()
@@ -21,3 +24,10 @@ class AsyncResponseIterator(object):
 
     async def __anext__(self):
         return await self._next()
+
+
+async def get_first_message_with_timeout(stream: AsyncResponseIterator, timeout: int):
+    async def get_first_response():
+        return await stream.next()
+
+    return await asyncio.wait_for(get_first_response(), timeout)
